+++ src/lib/messaging_handler.js
@@ -17,16 +17,76 @@ limitations under the License.
 
 "use strict";
 
-var env = process.env,
-  util = require('util'),
-  utils = require('./utils.js');
+var _ = require('lodash');
+var env = process.env;
+var util = require('util');
+var utils = require('./utils.js');
+var truncate = require('truncate');
+
+/*
+  DefaultMessagingHandler.
+*/
+function DefaultMessagingHandler(robot, formatter) {
+  this.robot = robot;
+  this.formatter = formatter;
+  this.truncate_length = env.ST2_MAX_MESSAGE_LENGTH;
+}
+
+DefaultMessagingHandler.prototype.postData = function(data) {
+  var recipient, split_message, formatted_message,
+      text = "";
+
+  if (data.whisper && data.user) {
+    recipient = data.user;
+  } else {
+    recipient = data.channel;
+    text = (data.user && !data.whisper) ? util.format('%s: ', data.user) : "";
+  }
+
+  recipient = this.formatter.formatRecepient(recipient);
+  text += this.formatter.formatData(data.message);
+
+  // Ignore the delimiter in the default formatter and just concat parts.
+  split_message = utils.splitMessage(text);
+  if (split_message.pretext && split_message.text) {
+    formatted_message = util.format("%s\n%s", split_message.pretext, split_message.text);
+  } else {
+    formatted_message = split_message.pretext || split_message.text;
+  }
+
+  this.robot.messageRoom.call(this.robot, recipient, formatted_message);
+};
+
+DefaultMessagingHandler.normalizeAddressee = function(msg) {
+  return {
+    room: msg.message.user.room,
+    name: msg.message.user.name
+  };
+};
+
+DefaultMessagingHandler.prototype.formatData = function(data) {
+  if (utils.isNull(data)) {
+    return "";
+  }
+  if (this.truncate_length > 0) {
+    data = truncate(data, this.truncate_length);
+  }
+  return data;
+};
+
+DefaultMessagingHandler.prototype.formatRecepient = function(recepient) {
+  return recepient;
+};
+
+DefaultMessagingHandler.prototype.normalizeCommand = function(command) {
+  return command;
+};
 
 /*
   SlackMessagingHandler.
 */
 function SlackMessagingHandler(robot, formatter) {
-  this.robot = robot;
-  this.formatter = formatter;
+  DefaultMessagingHandler.call(this, robot, formatter);
 
   var sendMessageRaw = function(message) {
     /*jshint validthis:true */
@@ -41,6 +101,8 @@ function SlackMessagingHandler(robot, formatter) {
     }
   }
 }
+
+util.inherits(SlackMessagingHandler, DefaultMessagingHandler);
 
 SlackMessagingHandler.prototype.postData = function(data) {
   var recipient, attachment_color, split_message,
@@ -93,20 +155,46 @@ SlackMessagingHandler.prototype.postData = function(data) {
   }
 };
 
-SlackMessagingHandler.normalizeAddressee = function(msg) {
+SlackMessagingHandler.prototype.normalizeAddressee = function(msg) {
   return {
     name: msg.message.user.name,
     room: msg.message.room
   };
 };
 
+SlackMessagingHandler.prototype.formatData = function(data) {
+  if (utils.isNull(data)) {
+    return "";
+  }
+  // For slack we do not truncate or format the result. This is because
+  // data is posted to slack as a message attachment.
+  return data;
+};
+
+SlackMessagingHandler.prototype.formatRecepient = function(recepient) {
+  return recepient;
+};
+
+SlackMessagingHandler.prototype.normalizeCommand = function(command) {
+  // replace left double quote with regular quote
+  command = command.replace(/\u201c/g, '\u0022');
+  // replace right double quote with regular quote
+  command = command.replace(/\u201d/g, '\u0022');
+  // replace left single quote with regular apostrophe
+  command = command.replace(/\u2018/g, '\u0027');
+  // replace right single quote with regular apostrophe
+  command = command.replace(/\u2019/g, '\u0027');
+  return command;
+};
+
 /*
   HipchatMessagingHandler.
 */
 function HipchatMessagingHandler(robot, formatter) {
-  this.robot = robot;
-  this.formatter = formatter;
+  DefaultMessagingHandler.call(this, robot, formatter);
 }
+
+util.inherits(HipchatMessagingHandler, DefaultMessagingHandler);
 
 HipchatMessagingHandler.prototype.postData = function(data) {
   var recipient, split_message, formatted_message,
@@ -143,7 +231,7 @@ HipchatMessagingHandler.prototype.postData = function(data) {
   }
 };
 
-SlackMessagingHandler.normalizeAddressee = function(msg) {
+HipchatMessagingHandler.prototype.normalizeAddressee = function(msg) {
   var name = msg.message.user.name;
   var room = msg.message.room;
   if (room === undefined) {
@@ -155,13 +243,34 @@ SlackMessagingHandler.normalizeAddressee = function(msg) {
   };
 };
 
+HipchatMessagingHandler.prototype.formatData = function(data) {
+  if (utils.isNull(data)) {
+    return "";
+  }
+  // HipChat has "show more" capability in messages so no truncation.
+  return '/code ' + data;
+};
+
+HipchatMessagingHandler.prototype.formatRecepient = function(recepient) {
+  var robot_name = env.HUBOT_HIPCHAT_JID.split("_")[0];
+  var hipchat_domain = (env.HUBOT_HIPCHAT_XMPP_DOMAIN === 'btf.hipchat.com') ?
+                       'conf.btf.hipchat.com' : 'conf.hipchat.com';
+  return util.format('%s_%s@%s', robot_name, recepient, hipchat_domain);
+};
+
+HipchatMessagingHandler.prototype.normalizeCommand = function(command) {
+  return command;
+};
+
+
 /*
   YammerMessagingHandler.
 */
 function YammerMessagingHandler(robot, formatter) {
-  this.robot = robot;
-  this.formatter = formatter;
+  DefaultMessagingHandler.call(this, robot, formatter);
 }
+
+util.inherits(YammerMessagingHandler, DefaultMessagingHandler);
 
 YammerMessagingHandler.prototype.postData = function(data) {
   var recipient, split_message, formatted_message,
@@ -188,50 +297,10 @@ YammerMessagingHandler.prototype.postData = function(data) {
   this.robot.send.call(this.robot, recipient, formatted_message);
 };
 
-YammerMessagingHandler.normalizeAddressee = function(msg) {
+YammerMessagingHandler.prototype.normalizeAddressee = function(msg) {
   return {
     name: String(msg.message.user.thread_id),
     room: msg.message.room
-  };
-};
-
-/*
-  DefaultMessagingHandler.
-*/
-function DefaultMessagingHandler(robot, formatter) {
-  this.robot = robot;
-  this.formatter = formatter;
-}
-
-DefaultMessagingHandler.prototype.postData = function(data) {
-  var recipient, split_message, formatted_message,
-      text = "";
-
-  if (data.whisper && data.user) {
-    recipient = data.user;
-  } else {
-    recipient = data.channel;
-    text = (data.user && !data.whisper) ? util.format('%s: ', data.user) : "";
-  }
-
-  recipient = this.formatter.formatRecepient(recipient);
-  text += this.formatter.formatData(data.message);
-
-  // Ignore the delimiter in the default formatter and just concat parts.
-  split_message = utils.splitMessage(text);
-  if (split_message.pretext && split_message.text) {
-    formatted_message = util.format("%s\n%s", split_message.pretext, split_message.text);
-  } else {
-    formatted_message = split_message.pretext || split_message.text;
-  }
-
-  this.robot.messageRoom.call(this.robot, recipient, formatted_message);
-};
-
-DefaultMessagingHandler.normalizeAddressee = function(msg) {
-  return {
-    room: msg.message.user.room,
-    name: msg.message.user.name
   };
 };
